move Abstract Test to specific dir
