+++ packages/NodeTraverserQueue/tests/BetterNodeFinderTest.php
@@ -7,7 +7,7 @@ use PhpParser\Node\Expr\Array_;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\ClassLike;
-use Rector\NodeTypeResolver\Tests\AbstractNodeTypeResolverTest;
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\AbstractNodeTypeResolverTest;
 
 final class BetterNodeFinderTest extends AbstractNodeTypeResolverTest
 {
