+++ packages/NodeTypeResolver/tests/PerNodeCallerTypeResolver/MethodCallCallerTypeResolver/NestedMethodCallTest.php
@@ -5,7 +5,7 @@ namespace Rector\NodeTypeResolver\Tests\PerNodeCallerTypeResolver\MethodCallCall
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Identifier;
 use Rector\Node\Attribute;
-use Rector\NodeTypeResolver\Tests\AbstractNodeTypeResolverTest;
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\AbstractNodeTypeResolverTest;
 
 final class NestedMethodCallTest extends AbstractNodeTypeResolverTest
 {
