+++ packages/NodeTypeResolver/tests/PerNodeTypeResolver/PropertyTypeResolver/PropertyTypeResolverTest.php
@@ -3,7 +3,7 @@
 namespace Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\PropertyTypeResolver;
 
 use PhpParser\Node\Stmt\Property;
-use Rector\NodeTypeResolver\Tests\AbstractNodeTypeResolverTest;
+use Rector\NodeTypeResolver\Tests\PerNodeTypeResolver\AbstractNodeTypeResolverTest;
 
 final class PropertyTypeResolverTest extends AbstractNodeTypeResolverTest
 {
